*** empty log message ***

+++ node/trunk/www/cron.php
@@ -39,7 +39,8 @@ sotf_Object::doUpdates();
 
 // sync with all neighbours
 $rpc = new rpc_Utils;
-$neighbours = sotf_Neighbour::getAll();
+$neighbours = sotf_Neighbour::listAll();
+debug("neihbours", $neighbours);
 if(count($neighbours) > 0) {
   while(list(,$neighbour) = each($neighbours)) {
       $neighbour->sync();
@@ -59,6 +60,7 @@ if(count($nodes) > 0) {
 }
 
 //********* IMPORT ARRIVED XBMF
+
 $dirPath = $config['xbmfInDir'];
 $dir = dir($dirPath);
 while($entry = $dir->read()) {
@@ -70,14 +72,16 @@ while($entry = $dir->read()) {
 	}
 }
 $dir->close();
-foreach($XBMF as $xbmfFile) {
-	$id = sotf_Programme::importXBMF($config['xbmfInDir'] . "/$xbmfFile", $config['publishXbmf']);
-	if($id) {
-		debug("CRON","Imported new XBMF: $xbmfFile");
-    unlink($config['xbmfInDir'] . "/$xbmfFile");
-	} else {
-    logger("CRON","Import FAILED for XBMF: $xbmfFile");
-	}
+if(!empty($XBMF)) {
+  foreach($XBMF as $xbmfFile) {
+    $id = sotf_Programme::importXBMF($config['xbmfInDir'] . "/$xbmfFile", $config['publishXbmf']);
+    if($id) {
+      debug("CRON","Imported new XBMF: $xbmfFile");
+      unlink($config['xbmfInDir'] . "/$xbmfFile");
+    } else {
+      logger("CRON","Import FAILED for XBMF: $xbmfFile");
+    }
+  }
 }
